Lab2/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pydantic[email]==2.9.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.0
openai==1.50.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
reportlab==4.2.2
pytest-cov==5.0.0
